requirements-dev.txt
PyQt6>=6.0.0
beautifulsoup4>=4.9.3
PyYAML>=6.0
lxml>=4.9.0
pillow>=9.0.0
pytest>=7.0.0
pyinstaller>=5.0.0

